+++ src/platform_impl/macos/monitor.rs
@@ -18,10 +18,13 @@ use crate::{
 };
 
 #[derive(Derivative)]
-#[derivative(Debug, Clone)]
+#[derivative(Debug, Clone, PartialEq)]
 pub struct VideoMode {
+    #[derivative(PartialEq = "ignore")]
     pub(crate) size: (u32, u32),
+    #[derivative(PartialEq = "ignore")]
     pub(crate) bit_depth: u16,
+    #[derivative(PartialEq = "ignore")]
     pub(crate) refresh_rate: u16,
     pub(crate) monitor: MonitorHandle,
     #[derivative(Debug = "ignore")]